Added the get_user_group function since the logic was being used in two places now. Might be overkill.
Updated the comments on the merge_files function to include ssh value

Added skipping error logic if the skip_on_user_group_error is set to true but the user or group doesn't exist in the environment.

In the main function, added a new merge_files phase which would be the ssh keys.