OPS-13631 Fix logic so that ef-open tools behave correctly on Jenkins and elsewhere (#133)
* Apply De Morgan's theorem so we can invert the symbols but maintain the same result. This helps make the condition easier to read than the old way on when pull_repo() occurs.

Also I don't think the translated logic was correct. We only want to pull repo when when the --devel flag is missing and we're not running on jenkins. So this translates to
if not CONTEXT.devel and CONTEXT.whereami != 'jenkins':

Fixed some comments and added messages so it's more obvious why the pull_repo() occurs.

* remove commented out code